pyproject.toml
[project]
name = "getpet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-minify>=0.50",
    "markdown>=3.9",
    "python-dotenv>=1.1.0",
    "tomli>=2.2.1",
]
